restructure how we establish our runtime and editor
