+++ .gitignore
@@ -1,5 +1,5 @@
 # Project Specific
-editor/bin/*
+bin/*
 
 # User Specific Generated Files
 **/.vs/*
