+++ CMakeLists.txt
@@ -4,6 +4,8 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_DEBUG_POSTFIX "-d")
 
+set(genesis_root_dir ${CMAKE_CURRENT_SOURCE_DIR})
+
 
 project(genesis VERSION "0.0.1")
 
@@ -36,7 +38,9 @@ if(PROJECT_NAME STREQUAL CMAKE_PROJECT_NAME)
   set(is_root_project ON)
 endif()
 
-option(GENESIS_BUILD_EDITOR "Build the genesis editor (else only library)" ON)
+option(GENESIS_ENABLE_EDITOR "Set the runtime to instead build the editor instead of the default" ON)
+option(GENESIS_BUILD_RUNTIME "Build the genesis runtime (else only library)" ON)
+option(GENESIS_BUILD_GAME "Build the genesis game" ON)
 option(GENESIS_BUILD_SHADERS "Build the genesis shaders" ON)
 option(GENESIS_BUILD_TOOLS "Build Genesis tools" ON)
 option(GENESIS_AUTOFORMAT "Run code-formatter before building Genesis" ON)
@@ -45,17 +49,33 @@ if(NOT GENESIS_BUILD_TOOLS AND GENESIS_AUTOFORMAT)
   message(FATAL_ERROR "Cannot enable GENESIS_AUTOFORMAT without GENESIS_BUILD_TOOLS")
 endif()
 
-if(GENESIS_BUILD_EDITOR)
-    set(GENESIS_BUILD_SHADERS ON)
+if(GENESIS_ENABLE_EDITOR)
+  set(GENESIS_BUILD_RUNTIME ON) # Editor requires the runtime lib
+endif()
+
+if(GENESIS_BUILD_RUNTIME)
+  set(GENESIS_BUILD_GAME ON) # Editor requires the game lib
+  set(GENESIS_BUILD_SHADERS ON)
 endif()
 
 
 add_subdirectory(ext)
 add_subdirectory(engine)
 
-if(GENESIS_BUILD_EDITOR)
-  add_subdirectory(editor)
-endif()
+if (GENESIS_BUILD_GAME)
+  add_subdirectory(game)
+endif ()
+
+# Currently we are only building the editor as our runtime, but eventually
+# we will have a separate runtime and editor with the editor runtime essentially hooking
+# a purpose built editor into the runtime.
+if (GENESIS_BUILD_RUNTIME)
+  if (GENESIS_ENABLE_EDITOR)
+    add_subdirectory(editor)
+  else ()
+    add_subdirectory(editor) # TODO: Replace with a default runtime.
+  endif ()
+endif ()
 
 if (GENESIS_BUILD_TOOLS)
   add_subdirectory(tools/code-formatter)
@@ -63,13 +83,13 @@ if (GENESIS_BUILD_TOOLS)
   if (GENESIS_AUTOFORMAT)
     add_custom_target(autoformat ALL
       code-formatter -q engine
-      COMMAND code-formatter -q editor/src
+      COMMAND code-formatter -q game/src
       WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
       USES_TERMINAL
       COMMENT "Formatting sources"
     )
 
     add_dependencies(genesis autoformat)
-    add_dependencies(genesis-editor autoformat)
+    add_dependencies(genesis-game autoformat)
   endif()
 endif()
