+++ editor/src/main.cpp
@@ -1,9 +1,9 @@
 // Copyright (c) 2023-present Genesis Engine contributors (see LICENSE.txt)
 
+#include <game/game.hpp>
 #include <gen/logger/instance.hpp>
 #include <gen/logger/log.hpp>
 #include <gen/util/version.hpp>
-#include "game.hpp"
 
 // TODO: Replace this with a config file. At least for the startup window size.
 static constexpr const char * appName{"Genesis Game"};
