+++ game/game_list.cmake
@@ -1,0 +1,4 @@
+
+set(game_headers
+        include/game/game.hpp
+)
