+++ game/src/CMakeLists.txt
@@ -1,3 +1,4 @@
+
 target_sources(${PROJECT_NAME} PRIVATE
-        main.cpp
+        game.cpp
 )
