+++ shaders/math/const.hlsl
